pyproject.toml
[tool.poetry]
name = "lr7"
version = "0.1.0"
description = ""
authors = ["Igorik <egorsborsikov250575@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
black = "^24.10.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
